requirements.txt
PyMuPDF==1.23.6
python-docx==1.1.0